apply RobinTF suggestions

+++ test/SerializerTest.cpp
@@ -324,7 +324,7 @@ auto testWithCallableSerializer = [](auto testFunction) {
 
   WriteViaCallableSerializer writer{write};
   auto makeReaderFromWriter = [&buffer]() {
-    auto read = [pos = uint64_t{0}, &buffer](char* target,
+    auto read = [pos = size_t{0}, &buffer](char* target,
                                              size_t numBytes) mutable {
       std::copy(buffer.begin() + pos, buffer.begin() + pos + numBytes, target);
       pos += numBytes;
