+++ test/SparqlExpressionTest.cpp
@@ -224,7 +224,7 @@ auto testNaryExpressionImpl = [](auto&& makeExpression, auto const& expected,
     if constexpr (isVectorResult<T>) {
       return operand.size();
     } else if constexpr (std::is_same_v<T, ad_utility::SetOfIntervals>) {
-      return operand._intervals.empty() ? uint64_t{0}
+      return operand._intervals.empty() ? size_t{0}
                                         : operand._intervals.back().second;
     }
     return 1;
